+++ src/briefy/leica/models/job/assignment.py
@@ -42,7 +42,7 @@ __listing_attributes__ = __summary_attributes__ + [
     'set_type', 'number_required_assets', 'category', 'availability', 'additional_compensation',
     'reason_additional_compensation', 'assignment_internal_qa', 'requirements', 'pool_id',
     'location', 'project', 'closed_on_date', 'pool', 'delivery', 'refused_times',
-    'asset_types', 'external_state'
+    'asset_types', 'external_state', 'current_type'
 ]
 
 __colander_alchemy_config_overrides__ = \
@@ -395,6 +395,9 @@ class Assignment(AssignmentDates, mixins.AssignmentRolesMixin, mixins.Assignment
 
     Instance of :class:`briefy.leica.models.project.Project`.
     """
+
+    current_type = association_proxy('order', 'current_type')
+    """Current type of the order."""
 
     location = orm.relationship(
         'OrderLocation',